@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = '1.3.72'
+    ext.kotlin_version = '1.5.31'
     repositories {
-        jcenter()
+        mavenCentral()
         google()
     }
     dependencies {
@@ -15,7 +15,7 @@ buildscript {
 
 allprojects {
     repositories {
-        jcenter()
+        mavenCentral()
         // required to find the project's artifacts
         maven { url 'https://maven.google.com' }
         maven { url 'https://jitpack.io' }
@@ -29,24 +29,24 @@ task clean(type: Delete) {
 
 ext {
     coroutines = '1.3.4'
-    corektx = '1.3.2'
-    koin_version= '2.1.5'
-    appcompat = '1.1.0'
-    constraint_layout = '1.1.3'
+    corektx = '1.6.0'
+    koin_version= '2.2.3'
+    appcompat = '1.3.1'
+    constraint_layout = '2.1.1'
     cardview = '1.0.0'
-    material = '1.3.0'
+    material = '1.4.0'
     preferencev14 = '1.0.0'
     appIntro = '6.1.0'
-    roomVersion = '2.2.6'
+    roomVersion = '2.3.0'
     archLifecycleExtensionsVersion = '2.2.0'
     archLifecycleVersion = '2.3.1'
-    nav_version = "2.3.4"
+    nav_version = "2.3.5"
     retrofit2 = '2.9.0'
     retrofit2gson = '2.9.0'
     leakCanary = '2.4'
     junit = '4.13'
-    extjunit = '1.1.1'
-    testrunner = '1.2.0'
+    extjunit = '1.1.3'
+    testrunner = '1.4.0'
     annotation = '1.1.0'
     debugdb = '1.0.6'
 }
