@@ -1,6 +1,5 @@
 package com.ocram.qichwadic.features.dictionaries.ui
 
-
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -9,7 +8,6 @@ import android.widget.AdapterView
 import android.widget.ArrayAdapter
 import androidx.appcompat.widget.Toolbar
 import androidx.drawerlayout.widget.DrawerLayout
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.AppBarConfiguration
@@ -20,22 +18,18 @@ import com.google.android.material.snackbar.Snackbar
 import com.ocram.qichwadic.R
 import com.ocram.qichwadic.core.domain.model.DictionaryModel
 import com.ocram.qichwadic.core.ui.DictLang
-import kotlinx.android.synthetic.main.fragment_dictionary.*
+import com.ocram.qichwadic.core.ui.fragment.BaseFragment
+import com.ocram.qichwadic.databinding.FragmentDictionaryBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import java.util.*
 
-class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, DictionaryAdapter.DefinitionDownloadListener {
+class DictionaryFragment : BaseFragment<FragmentDictionaryBinding>(), AdapterView.OnItemSelectedListener, DictionaryAdapter.DefinitionDownloadListener {
 
     private val dictionaryViewModel: DictionaryViewModel by viewModel()
     private lateinit var dictionaryAdapter: DictionaryAdapter
     private lateinit var dictLangMap: Map<String, List<DictionaryModel>>
     private lateinit var dictLangs: List<DictLang>
     private var currentLangCode: String? = null
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_dictionary, container, false)
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         val navController = findNavController()
@@ -47,10 +41,10 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
         toolbar.setupWithNavController(navController, appBarConfiguration)
 
         initViews()
-        dictionaryViewModel.dictionariesByLang.observe(this, Observer { this.onDictionaryListChanged(it) })
-        dictionaryViewModel.localLoading.observe(this, Observer { this.onLocalDictionariesLoadingChanged(it) })
-        dictionaryViewModel.dictionaryActionStatus.observe(this, Observer { this.onDictionaryAction(it) })
-        dictionaryViewModel.cloudError.observe(this, Observer {  this.onCloudError(it) })
+        dictionaryViewModel.dictionariesByLang.observe(viewLifecycleOwner, Observer { this.onDictionaryListChanged(it) })
+        dictionaryViewModel.localLoading.observe(viewLifecycleOwner, Observer { this.onLocalDictionariesLoadingChanged(it) })
+        dictionaryViewModel.dictionaryActionStatus.observe(viewLifecycleOwner, Observer { this.onDictionaryAction(it) })
+        dictionaryViewModel.cloudError.observe(viewLifecycleOwner, Observer {  this.onCloudError(it) })
     }
 
     private fun initViews() {
@@ -58,26 +52,26 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
 
         setRecyclerView()
         setSpinnerAdapters()
-        spTargetLanguages.onItemSelectedListener = this
+        binding.spTargetLanguages.onItemSelectedListener = this
     }
 
     private fun setRecyclerView() {
         dictionaryAdapter = DictionaryAdapter(ArrayList(), this)
         val linearLayoutManager = LinearLayoutManager(requireContext(), RecyclerView.VERTICAL, false)
-        rvDictionaries.layoutManager = linearLayoutManager
-        rvDictionaries.isNestedScrollingEnabled = false
-        rvDictionaries.adapter = dictionaryAdapter
+        binding.rvDictionaries.layoutManager = linearLayoutManager
+        binding.rvDictionaries.isNestedScrollingEnabled = false
+        binding.rvDictionaries.adapter = dictionaryAdapter
     }
 
     private fun setSpinnerAdapters() {
         val langAdapter = ArrayAdapter(requireContext(), R.layout.item_spinner_dict_lang_white, dictLangs)
         langAdapter.setDropDownViewResource(R.layout.item_spinner_dict_lang_white)
-        spTargetLanguages.adapter = langAdapter
-        spTargetLanguages.setSelection(0)
+        binding.spTargetLanguages.adapter = langAdapter
+        binding.spTargetLanguages.setSelection(0)
     }
 
     override fun onItemSelected(p0: AdapterView<*>?, p1: View?, pos: Int, p3: Long) {
-        val dictLang = spTargetLanguages.getItemAtPosition(pos) as DictLang
+        val dictLang = binding.spTargetLanguages.getItemAtPosition(pos) as DictLang
         this.currentLangCode = dictLang.code
         dictionaryAdapter.dictionaries = dictLangMap[dictLang.code]
         dictionaryAdapter.notifyDataSetChanged()
@@ -103,13 +97,13 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
     private fun showDictionaries(dictionaries: List<DictionaryModel>?) {
         dictionaryAdapter.dictionaries = dictionaries
         dictionaryAdapter.notifyDataSetChanged()
-        spTargetLanguages.visibility = View.VISIBLE
-        rvDictionaries.visibility = View.VISIBLE
+        binding.spTargetLanguages.visibility = View.VISIBLE
+        binding.rvDictionaries.visibility = View.VISIBLE
     }
 
     private fun onLocalDictionariesLoadingChanged(isLoading: Boolean?) {
         if (isLoading != null) {
-            pbDictionariesLoading.visibility = if (isLoading) View.VISIBLE else View.GONE
+            binding.pbDictionariesLoading.visibility = if (isLoading) View.VISIBLE else View.GONE
         }
     }
 
@@ -141,7 +135,14 @@ class DictionaryFragment : Fragment(), AdapterView.OnItemSelectedListener, Dicti
 
     private fun showMessage(message: String?) {
         if (!message.isNullOrEmpty()) {
-            Snackbar.make(clDictionaries, message, Snackbar.LENGTH_SHORT).show()
+            Snackbar.make(binding.clDictionaries, message, Snackbar.LENGTH_SHORT).show()
         }
+    }
+
+    override fun viewBindingClass(
+        inflater: LayoutInflater,
+        container: ViewGroup?
+    ): FragmentDictionaryBinding {
+        return FragmentDictionaryBinding.inflate(inflater, container, false)
     }
 }
