@@ -6,10 +6,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ArrayAdapter
-import android.widget.TextView
 
 import com.ocram.qichwadic.R
 import com.ocram.qichwadic.core.ui.DictLang
+import com.ocram.qichwadic.databinding.ItemSpinnerDictLangBinding
+import com.ocram.qichwadic.databinding.ItemSpinnerDropdownSearchTypeBinding
+import java.util.*
 
 class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, R.layout.item_spinner_dict_lang) {
 
@@ -34,24 +36,23 @@ class SearchDictLangAdapter(context: Context) : ArrayAdapter<DictLang>(context, 
         return -1
     }
 
-    override fun getItem(position: Int): DictLang? {
+    override fun getItem(position: Int): DictLang {
         return dictLangs[position]
     }
 
     @SuppressLint("DefaultLocale")
     private fun getCustomView(position: Int, parent: ViewGroup, dropdown: Boolean): View {
-        val inflater = LayoutInflater.from(parent.context)
-
-        val layoutId = if (dropdown) R.layout.item_spinner_dropdown_search_type else R.layout.item_spinner_dict_lang
-        val layout = inflater.inflate(layoutId, parent, false)
-
         val dictLang = getItem(position)
-        val tvName = layout.findViewById<TextView>(R.id.tvName)
-        if (dictLang != null) {
-            val displayText = if (dropdown) dictLang.name else dictLang.code.toUpperCase()
-            tvName.text = displayText
+        val inflater = LayoutInflater.from(context)
+        return if (dropdown) {
+            val binding = ItemSpinnerDropdownSearchTypeBinding.inflate(inflater, parent, false)
+            binding.tvName.text = dictLang.name
+            binding.root
+        } else {
+            val binding = ItemSpinnerDictLangBinding.inflate(inflater, parent, false)
+            binding.tvName.text = dictLang.code.uppercase(Locale.getDefault())
+            binding.root
         }
-        return layout
     }
 
     override fun getDropDownView(position: Int, convertView: View?, parent: ViewGroup): View {
